[BC-BREAKING] Turn off scalar_check for masked_select.
Note that this changes the behavior of masked_select when both "self" and "mask" are 0-dimensional.

In previous versions of PyTorch, this would return a 0-dimensional tensor.  But the documentation reads:
"Returns a new 1-D tensor which indexes the input tensor according to the boolean mask mask which is a BoolTensor."

ghstack-source-id: ef34fc69bcbabe62268372b08ace585fb074d0ce
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29923