Fix type sharing on loaded ScriptModules
After save/load, we lose concrete type information. So if you tried to
script something that contained a loaded ScriptModule as a submodule,
the following sequence happened:
1. During ConcreteType inference, the loaded submodule got a new
inferred type.
2. But it already has a type! So there was a type mismatch.

To fix this, we should generate a ConcreteType directly from the loaded
submodule type (similar to what we do for interfaces). This makes sense
too--the ConcreteModuleType should be empty, since all the "sugaredness"
was stripped out during the save/load process.

ghstack-source-id: 6763d3a3bde283fe1ddae450693fdcb7db39f944
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29826